requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
mysqlclient==2.2.4
nltk==3.8.1
pyasn1==0.6.0
pycryptodome==3.20.0
regex==2024.5.15
rsa==4.9
tinyec==0.4.0
tqdm==4.66.4
Werkzeug==3.0.3
zipp==3.19.2
